requirements.txt
# AkMon 项目依赖
akshare>=1.12.0
pydantic>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.0.0
pytest-cov>=4.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
